Mega65: ask Xemu user to answer kick upgrade
This needs SDL 'chooser' window to be used.
The intent of this feature: use may have newer KICKUP.M65
than on the SD card image, and *don't* want to be upgraded,
also it can be useful to test upgrade/not upgrade paths as
well, without bothering to modify the SD card image again
and again to have or not have the KICKUP.M65 file on it,
and what version exactly.