Mega65: valid code points check fix
Code points means here: information from kickstart.list
about PC locations where there is really code.

Previous verson had a type on code, so this feature not
worked for real. Now it does. Moreover, on kickstart
upgrade from SD, this feature is disabled, as it's quite
likely, the other kickstart on the card is not the very
same binary with the same "code points".