Mega65: hack for "32 bit STA (ZP),Z" GS op
Kickstart seems to use this opcode. A real M65 has this for
multiple opcodes not just this, but it's only a quick hack.
Also, writing to the "VIC-IV charrom write area" (whatever
it is ...) is ignored to avoid a segfault :) Now kickstart
returns from hypervisor mode at the end to run ROM code :)
However there is some problem DMA is left to be confused as
it wants to access >127Mbyte area by the ROM code already.