Mega65: internal kickstart for poor men :)
If Xemu cannot find the KICKUP.M65, it can use its
internal version, thus now require only the SD-card
image at the worst case, no absolute need for
KICKUP.M65 "externally", neither the C65 ROM images.
It's more like a real M65, where it's enough to
"feed" it with an SD-card with the right files on it
though.