Core: snapshot framework #39
Trying to create an emulation target independent snapshot
format and common routines to read and write them with the
need of implementation by the targets. Currently it does
not do anything useful. With the C65 emulator some can try
to load/save snapshots (hint: command line option) but it
merely saves/load memory only, so it won't work as it's
expected. Also, it can be still buggy and ugly ...