M65: quick hack for unaligned SD access keep busy
Idea from Gurce. Though it must be noted, that it seems new
KS does SDHC detection by inspecting unalinged SD card access
and wathing only the BUSY flags, not even the error ones. This
sounds strange for me, but anyway, it seems to work now with
newer KS as well ...