M65: allow to force external loaded ROM
Xemu/M65 contains a "preinit memory cache" which can be filled with
different -load* options, like -loadrom for the C65 ROM. However
kickstart usually loads ROM from the SD-card image, overwriting the
pre-init'ed memory content. -forcerom option is introduced by this
commit, which re-copies the ROM from the pre-init cache to the memory on
the fist hypervisor leave call, which should be the point when
hypervisor transfers control to the ROM image, thus the right time to
re-init the content. -forcerom should be used with pair of -loadrom FILE
for the function, though -loadrom itself (without -forcerom) can be
useful for other purposes, thus the reason for non-default
interpretation of this function. Issue #69