Mega65: sdcard further cleanup + bugfixes
It seems, the Hyppo used with Xemu/Mega65 still probes SDHC.
Since I killed non-SDHC support, that's a problem, as Xemu
tries to handle silently everything as SDHC so the probe
will be bad. Hack-in some ugly solution to FAIL ALL I/O
operations if non-SDHC mode is selected. It seems it cures
the problem. Also virtual-storage had a design flaw with
all-zeroes-block write, fixed. Some misc beauty work was
done, and some dead code is eliminated. Though, still
work-in-progress!!