M65: auto-format just created SDcard image by Xemu
Xemu provides auto creation of the SD-card image file if it's the
default one (no user override on its path) and it haven't existed yet.
Thus, it's logical to assume, the user wants to format it. Since it has
been just created, really, no danger to format it (no data loss) thus
it's even not worth to bother the user with questions if it's OK to do
so, or ask the user to use the UI menu to do it. This way, many
confusion can be avoided for new users, who sees Hyppo's "menu" and try
to use that to format (which works of course, but more complicated and
slower). Also the final Xemu signature check has been modified, to tell
the user how to access the menu at all, and he needs to update the files
on the card now.