C65+M65: F011 disk change signal problem #118
As Snoopy pointed out, disk change signal is bit 0 in status register B,
and not bit 7! That can be a reason of problems described in #118
However I'm still not sure if it's the only issue, as MEGA65 itself has
very similar problem, though there, it seems to be correct bit is used.