MEGA65: try to fix DMA for sticky options problem
It seems an enhanced mode DMA's "megabyte slice selection" option
remains active even for future DMA jobs, which is especially dangerous
if it's a C65 DMA (without any option to set "slice"). Reported by Bit
Shifter on Discord, with his new "EDMA" BASIC 10 (errr, BASIC 65)
command implementation.

This is only a try to fix it, really, the whole DMA emulation
must/should be rewritten as of #198 ...