UI: warn user about missing menu support #240
User may have been confused when right-clicking into the emulator window
according to the documentation but nothing happens. This can happen
because no UI support is compiled in which is capable of displaying
menus. This can happen more on UNIX platforms when failing to compile
GTK3 support. To somewhat ease the confusion, let's drop some error
window to explain the situation for the user a bit. Well, better than
nothing.