UI/OSD: re-structure OSD source #224 #203
OSD has been separated to emultools.c (and .h) included from xemu/gui/
directory. This way it's much easier to maintain and develop further.
XEMU_OSD_SUPPORT macro was added, needed to be set in xemu-target.h of
emulators to declare the need of the OSD support. Some variable
reorganization happened as well.