MEGA65: try to fix div_mult unit div result #229
Divide is done differently than I thought. It's not quotient and reminder,
but something like: ((Uint64)input_a << 32) / (Uint64)input_b as a 64 bit
result. Also, by inspecting VHDL, it seems if some tries to divide by zero,
the policy "do not update the result" is not true, but rather than setting
every bits to 1 in the result.