MEGA65: check NOP presence after hypervisor trap
On real MEGA65 there is a problem that after writing $D640-$D67F
(causing a hypervisor trap) on return, the next opcode _may_ not be
executed. Since it's not so deterministic actual problem, it's a must to
use a NOP, thus it's not a problem if executed but neither when it is
not. Xemu does not have a "non-deterministic" bug like this, however
this can be dangerous to develop a problem with Xemu which won't work
with a real MEGA65 then. Thus, we want a warning in situations when the
NOP is missing.