M65+C65: CIA output read back fix try #250
Accordint to CIA datasheet, CIA for real always reads the value "on the
pins" regardless of DDR input or output. There was a forgotten callback
in Xemu always producing $FF for reading back CIA2 port A. Also, for
port B, it's interesting as it was also used on C65 to auto-run some
internal-to-ROM programs if certains pins are grounded. I tried to
approximate things here: for port A, the output value is read back,
other than the two high bits since those part of the IEC bus and in an
inverted manner corresponding to the lower-by-two two bits (which are
output). Currently it seems ROMs does not hang at least on boot, which
was a problem at the beginng to messing up the IEC bus state badly.