Adding initial video standard. See details below:
It seems, PAL is the de-facto default for MEGA65. Newer Hyppo's maybe
even set this, no problem. Also, it seems most user use MEGA65 as PAL.
Thus we have a problem that people assuming Xemu to be same as MEGA65
start to code and not realizing it's in NTSC (Xemu's Hyppo's
default). Till we are mature enough to be able to run a newer Hyppo, a
workaround: new option has been added to Xemu, -initvideostd which
sets the video mode after boot when hypervisor returns to execture ROM
code, by effectively overriding Hyppo's setting on "boot". This option
now defaults to PAL, so even without specifying this option. The old
-videostd switch has been renamed to -forcevideostd which is a different
thing: it forces a video mode, and you cannot even alter it ever later,
not just the initial state (sure, using -forcevideostd will render
-initvideostd somewhat useless, no need to use both!).