Fix transparency at some places #29
16 colour sprites: if I remember correctly, in 16 colour sprites mode,
the sprite colour VIC registers gives the palette index which is
transparent, not the palette index zero is the transparent one in all
cases.

Also, I modified some places where the bg buffer was filled with SDL
pixel value. Also, some meaningless cosmetic changes :)

On the long term, I would suggest, to re-construct that bg buffer has
meaning as zero meaning "bg" and any other value as "fg", rather than
just the two constants. It will improve performance as well since in
some cases then it's enough to push value as-is (without a condition) if
anyway zero value is already means bg and other value as fg. Surely this
does not apply at every cases, but in some cases it does, the other
cases - at the other hand - won't be affected badly in performance
either, so nothing to lose with this proposed change.