Fixing MCM sprite rendering
Warning, I haven't tested this! MCM sprite rendering was incorrect,
since it checked "pixel" to be transparent, but pixel was decoded to the
palette entry not the actual two-mcm-bit pixel value. Also lookup table
is more nice/faster. This is my version which may be also bad, but
who knows ;) For #29