MEGA65: do not interrupt prefixed opcodes at 1MHz
MEGA65 old behaviour was to allow interrupts after the prefix if 1 or
2MHz was used. Thus using compound opcodes were dangerous at slow
speeds, if interrupts were not disabled. However it seems this has been
changed since then (even the manual needed to be updated to reflect this
change). This rudimentary fix simply checks the prefix. Probably not the
very same what MEGA65 does though ...