MEGA65: try to fix new ROM doesn't load issue
Long standing problem: after updating ROM files on the SD-card, for some
reason, Hyppo does not load the new version from the SD-card even after
reboot, and an Xemu total restart is needed. I try to workaround this
here, by copying the new ROM into the memory as well. If Hyppo does not
want to do it, I force it (evil laugh) ;)