MEGA65: use ext default.d81 if found
Common user complain: the default used "drive 8" disk image is inside
the SD-card image, thus very hard to deal with. This commit tries to
ease that: if no -8 option was used _AND_ a file default.d81 can be
found (with the right size of a D81) in the SDL-pref-dir, then it's
used by default. The user however still needs to put there a valid
D81 there, though.