D81ACCESS: fake "D64" mode
This a new "cheating" mode of the d81access layer (used to feed the disk
emulation of - currently - C65 and MEGA65's FDC). It just "converts" a
D64 to D81 "on-the-fly" at the time of the sector level acceses. It does
NOT give any compatibility with any 5.25" drive! It's just a comfort
service so user does not need to convert D64 to D81 to try to load
something from, nothing more. It's only the framework, currently it's
not used (it is, in the hyppo branch, already).

Also some cleanup, deletion of long ago removed code.