MEGA65: partial support for D64/D71 mount
This support uses D81 disk geometry still (as with MEGA65 for similar
feature!) but allow to access D64 or D71 (D65 is not yet
tested/supported). It needs the translation on geometry by software, eg
what closed-ROMs does recently.

Note, that it's a _partial_ support for now, won't work on SD-card
images, but only on external mounts.