M65: try to avoid window resize on power-on/reset
Xemu has to resize window size to match the texture if the texture
size changes. This happens on video standard change (PAL/NTSC). During
"power on" (start of the emulator) and reset, there is a transition
between PAL and NTSC causes an ugly glitch for a while. Trying to cure
this by disabling the ability to change video standard during the reset
trap to avoid this. Kinda a cosmetic change ...