CORE: SDL getbasepath fix, argc/v reorg.
SDL on OpenBSD seems to have problem with SDL_GetBasePath(). So as a
workaround I've put some code to use the absolute path from argv[0] or
the getcwd() result + relative path from argv[0] in this case.
Meanwhile, I had to realize that it needs to passing argv and argc to
xemu_pre_init() function anyway, so I can simplify some logic needed to
pass those again for the xemucfg_parse_all(). However this also means,
that I had to modify every emulators according to this API. I also put
some slight modification into MEGA65 emulator's matrix mode so things
can be queried to be able to check.

Thanks to rt.fm on Discord for testing Xemu on OpenBSD + spotting this
SDL2 bug.