PC: Import of unfinished PC emulator from 'dev'
Unfinished PC emulator using (reworked) FAKE86 infrastrcture. Please do
not even have a look on it right now.

The only reason of this commit to have some "base" for further
development on this emulator (outside of the 'dev' branch as well).

Note, that it should have unified with my tree of Fake86 (fork of the original
Fake86). Both of them may contain useful modifications/cleanups on the CPU
emulation which should be important to make nicer first ...

My Fake86 fork: https://github.com/lgblgblgb/fake86