CORE: better nameing last time funcs + new uts now
There are some functions in Xemu core, with very misleading names,
suggesting it returns with current time. However it returns that info
from the _last_ queried value used for timekeeping! To remedy this, some
functions now are decorated as "_last" as a postfix in their names, also
a new function is introduced to query Unix timestamp (as Uint64) which
really does what you expect.