MacOS: New Apple hw doesn't have right click? #358
It seems newer Apple notebooks do not have ability to produce right
click with the built-in touch pad (works with external mode). This is
shocking how lame Apple is, not having INS key, what Apple has at all,
removing all features? :-O Anyway ;) It seems the standard way is to use
CTRL + click to emulate right click. However that is not enforced by the
OS itself and must be done by the applications. SDL2 has this ability,
but must be requested. So this patch does exactly that.