Windows: trying to avoid segfault on console magic
The Windows Console Saga continues, and it is raging louder than ever ;)
Still far from being perfect, or could be understood, Windows is a total
mistery.

... now seriously. Thanks to Mirage_BD for the bug report and
investigation on an issue, causes Xemu to segfault on Windows if started
from Makefile. This patch wants to fix that, but even in just that it's
not perfect, now if run from Makefile, if you try to close/open syscon
(from UI) then you can't exit from Xemu since console window does not
want to be closed (but at least no segfault ...).