OSX: do not set min-osx-ver by default on build
Setting -mmacosx-version-min=X.Y on OSX dictates the minimal OSX version
the result should be OK on. However, if you build Xemu as your own, your
goal is 99.99% is to run Xemu on your machine. So setting this option is
not recommended, unless it's part of the official build process. The
problem: newer SDLs have stricter policy on the minimal OSX version and
people had problems to compile Xemu by their own.

So this commit deletes this option for the normal build/configure
process in general, and only from the .travis.yml file it's requested
(for official builds on Travis) via make option.