MEGA65: use only 3 bits for trimming in VIC-IV emu
MegaPoly intro does display "emptyness" on MEGA65. Mirage_BD (author of
MegaPoly) pointed out that maybe bit 3 of trim info does nothing on real
MEGA65 and should be ignored, though that bit is mentioned here and
there. But currently at least it SEEMS MEGA65 works this way, and it
also fixes MegaPoly to work correctly on Xemu as well.

On longer term, it can be the case that it's buggy even on a real
MEGA65 and this not the desired behaviour. However for now, let's follow
what MEGA65 _actually_ does, rather than what documentations/etc say.

Discussion link (on Discord):

https://discord.com/channels/719326990221574164/1043843197803175976/1113091458556252243