MEGA65: char "WOM" is readable now #376
The character-WOM (also called just the WOM) is called WOM because it was
Write-Only-Memory. It could be written by the CPU (only), and VIC-IV could
(obviously) only read. However the problem was: freezer could not save
the content of the "WOM" since it's not readable, and if user write into
it, the state won't be intact anymore. Basically the mega65-core change
makes the "WOM" (which shouldn't be called WOM anymore!! But I guess the
name will survive ...) readable, thus freezer can save its state (and then
restore, if needed).