CORE/C65/M65: add cfg-core new flag, + source-ids
Adding new flag into the config framework of Xemu to signal 'no-save' of
a given config option. It's used not allow to save that option when
configuration is directed to be written out, ie when user 'saves the
configuration'. This is desired, as some options more meant to be used
from CLI/scripts/tests etc, but if the user saves the config then, those
options are also saved. Typically the -prg is such an option, a user
doesn't want to be saved, 99.99% of the time (and if so, and used for
some testing, the user want to write a custom config file by hand anyway
as a special use case).

Also, this change does some other modifications in the config framework,
like adding "source" info of a config item, ie being CLI, config file or
"run-time". Currently it's not used too much.

Some new functions are added to deal with the functionalities above.

And finally, for now, MEGA65 and C65 targets uses the first mentioned
new thing.