MEGA65: fast boot mode
With the -fastboot switch, Xemu uses the "sleepless" timing during the
RESET HYPPO trap to speed up booting. Can be useful if called from
scripts with parameters like -prg and such to make the experience
faster. It's usless if -sleepless is used anyway, since then the
sleepless timing is always applied.