MEGA65: new compressed R/O SD image format
Can be more useful with emscripten #96 where a 4Gbyte sized image
is very unrealistic. The old format allowed compression of my image
to about 20-30Mbyte, this one is down to about 500K which means much
less data transfer needs from the browser. It uses RLE kind of
compression methods, and uncompressed on-the-fly per 64K "pages" basis
to have a balance between peformance, memory usage and compression.