MEGA65: matrix 2nd page + new disassembly ability
The second page feature is not perfect yet, causing glitches on paging
back and forth. Page UP/DOWN can be used to see the previous and the
current (active) page of the "matrix console".

Disassembly can be done in matrix mode with the "a" command, which
requires some address parameter, the syntax follows the "d" (dump)
command in that regard. Command "a" alone just continues the disassembly
from the last address.

WARNING: disassembly can be buggy/incorrect, it was not highly checked
yet, if it's 100% accurate. It knows the MEGA65 "compound" opcodes as
well.