ANDROID: porting to android: phase-1 #98
Currently nothing works yet, just the ability to actually run configure
for Android cross-compilation on UNIX, and allow the compilation
succeeded till the point of getting a shared object. Surely, this alone
is not useful for anything, an APK should be built, JNI activation
should be added, and these all do not mean it actually runs on any
Android device yet, still ...

Thanks btoschi on Discord for the chat on the topic and for some useful
tips here and there.