MEGA65: text pasting via hw kbd queue #391
Originally part of issue #103

Implementing some "paste text" (from OS paste buffer) is something which
is really hard to do with Xemu/MEGA65. A C64 emulator usually injects
into the software keyboard buffer implemented by the ROM. I cannot do
that with a MEGA65 emulator since the wide variation of ROMs and ROM
versions are in case. This method though used the _hardware_ keyboard
scanner queue which is not ROM dependent, so it's fine. However the
disadvantage: only works with newer ROMs and won't work in "GO64" mode.