CPU65: Opcode BIT #nn should work as all BIT ops!
For the CPU65 emulator I used (maybe old) datasheet on 65C02, which all stated
that BIT #immed is special, and only the Z flag is affected, unlike other BIT
addressing modes. However it seems it was just for early 65C02s. Also
mega65-core (and it seems also C65 iself) uses the regular implementation for
immediate addressing mode as well. Just in case, I left open the possibility
to use the old impementation if CPU65_OLD_BIT_IMMEDIATE is defined.