M65/CORE: allow winpos set/restore #383
This commit allows to set Xemu window position via CLI switch. Also, UI
menu has been added to be able to save current window position which is
restored on the next execution of the emulator. The save itself is NOT
automatic, since it seems, this is a bit awkward (not because of Xemu):
it seems sometimes SDL2 sees window pos _without_ the decoration border,
but sometimes with it ... So always saving winpos (window position) and
always restoring it can cause the window to draft away a bit at every
executuion, which is annoying ...