MEGA65: avoid ROM start lockup b/c CIA state #395
Note: this is _not_ a proper fix, rather than a workaround by setting
CIA latched values to $0101 initially on start-up. Without this, newer
beta ROMs (v920391 and v920392) is left in an infinite loop at their
start-up routines with some CIA checking magic ... As a real MEGA65
works without workarounds, this must be an Xemu issue, but the proper
fix is hard to develop because the inprecise basic foundation of Xemu
for clock-accurate CIA emulation :( See the mentioned #395 issue above
for more details.