BUILD: use newer ubuntu LTS on Travis CI
Travis has problems in the past with Ubuntu 20.04. I've issued a bug
report, and for a while no successfull builds could be done, so I used
Ubuntu 18.04 for a long time. Now it seems my bug report was addressed
by Travis, and it works now. It means that a newer mingw compiler is
used to produce Windows builds, which hopefully means a bit better
emulation performance on Windows. Or not ... ;)