BUILD: do not test grep -o on non-Linux #396
configure tests for working `grep` utility for building, however it
seems some systems (it seems NetBSD in particular) have problems with
the `-o` switch combined with others. As the `-o` is only used for real
at Linux hosts, let's not test `-o` if the build system is not Linux.

Reported by @Rhialto