CORE: simple debug hooks into CPU65 emu #11
To allow easier debugger implementation with breakpoints, and such, a
new (optional to be used) mechanism is introduced. It allows the CPU65
emulation core to call target emulator defined functions on each op code
executions (and in case of IRQ and NMI acceptance). The target emulator
itself can decide what to do in the callbacks, it can simply return, or
instructing the CPU emulation to be stopped (ie: breakpoint or step
based opcode execution during debugging).

NOTE: still, the target emulator's main loop must aware of this, the
CPU65 core itself alone cannot stop the whole emulation in a sane way.