MEGA65: first try of line drawing DMA (WIP!!) #404
Incomplete and possible buggy first try implementation of DMA line
drawing. Ideas are based on a "draft implementation" by btoschi
on Discord (@bjotos on github).

The only reason I present this highly incomplete work in the next branch
of Xemu (so not in dev) because I am on holiday, so it's a bit random
when I have time to do more work. Thus it's better to release it in
this unfinished form, I guess, if someone wants to play with it.