MEGA65: fix joy port swap with mouse grab #415
Also adding an config option '-joyport N' (N can be 1 or 2) as a CLI
option to select the default emulated port. Also, now the configuration
can be saved which preserves the "swap state" of ports.

The original problem was discovered by @Gurcei when he noticed that
using the UI menu's "swap port" functionalty does nothing. It was
because for some reason I wrote code to force port-1 when in mouse grab
mode. Which is bad ...