MEGA65: adding option to disable mouse emulation
The problem: Xemu user may leave mouse grab mode to do something with
their mouse. To avoid bothering the mouse-aware MEGA65 program running,
I returned zero for relative mouse position change in this case. However
that can cause problems with certain programs which are not mouse based
and expecting $FF if mouse is not there.