BUILD: Travis problem, Travis solution
The last build errored out on the Linux based builts. Travis response
was:

Thanks for reaching out and very sorry for this issue. We are aware of this
issue. As a workaround could you please add in your .travis.yml in the
deploy section "dpl_version: 2.0.5.2.beta.1".

Now, I am trying to do just like that. Fingers crossed ;)