MEGA65: H640 mustn't limit used bitplanes #421
Probably because of C65 limitation, MEGA65 emulation still had the limit
of used bitplanes to max 4 when H640 feature is used. I guess it was an
accident after the C65/MEGA65 emulation target split in Xemu near the
year of 2021 or so. However it results corrupted screen now when more
than 4 bitplanes (and/or other than the first 4)  are used with H640 bit
set.