MEGA65: HDOS setname should use Y register only
As discussed with FPhoton and Gurce on Discord, HDOS setname uses only Y
for the filename address transfer and must be page aligned (so Y
contains the high byte of the address). In the old days there was a
stage (IIRC) when X was used as well (as low byte, no page alignment was
needed then). The MEGA65 ROM still sets the X to zero when calling HDOS
probably to ensure compatibility, however newer software may choke on
this, not setting X register at all.

Discord thread:

https://discord.com/channels/719326990221574164/1314179483498188830/1339916222325587978