MEGA65: preliminary VIC4 CHARY16 (TCR) impl. #424
CHARY16 (or TCR meaning Tall ChaRacter) is a (so far relatively unknown)
mega65-core feature has become popular recently :)

It works, by fetching glyph information shifted by 2048 at every second
raster in V200 mode instead of the same. So it needs a special layout for
the chargen info if the mode is enabled (bit 4 of $D07A):

GS $D07A.4 VIC-IV:CHARY16 Alternate char ROM bank on alternate raster lines in V200

This is a bit problematic for Xemu, as it has an (otherwise not so correct)
infrastructure that in V200 it renders only every second line and basically
"clone" the previous one. This must be changed at some point for sure, but
now, probably an ugly hack will be done, to take an exception of that rule
if this mode is enabled.

**THIS IS AN UGLY HACK** Probably the longer term solution would be to
render all scanlines in all modes, this is already a problem in other
contexts as well!