MEGA65: on sysfile upgrade, update hdos dir too
With HDOS virtualization it can be a problem, that system files
(includign the ROM) cannot be seen anymore through Hyppo/HDOS functions,
since the SD-card image is not used then. For the emulator itself, it's
not a problem as Xemu loads the ROM directly, also Hyppo RESET trap
always uses the SD-card image (regardless of HDOS virtualization).
However it can be a problem, if a user program want to load the ROM (eg
it may destroy the ROM area after turning off ROM protection and want to
restore the ROM at the end).

To try to remedy this, UI menu "Update files on SD image" (in menu
"Disks / Cart" >> "SD-card") and also initial setup (for new Xemu
users) will also populate the "hdos" directory with the mentioned system
files.