MEGA65: MAP opcode behaviour change #427
As it turned out mega65-core modified the exact behaviour of the MAP
opcode in hypervisor mode back in 2022 with this commit:

https://github.com/MEGA65/mega65-core/commit/34a930b1a25988fdab4d22ae8f786b35e07799b7

See mega65-core issue: https://github.com/MEGA65/mega65-core/issues/635

Though Xemu still used the old behaviour which can be problematic.

So this is the intended fix to handle the problem.

Thanks @gurcei to discover / report this anomaly.