MEGA65: fix the fix for #429
My previous try to fix issue #429 was problematic:

Commit a44b71e35d40a5c19a420db8831bff112212bc78

This patch now is from FPhoton. Thanks.

Though I didn't want to introduce checking 'DEN' bit of $D011 inside a
renderer, my second solution would be just too complicated compared to
the scope of the problem which can be more easily fixed. So the usual
business, trivial to fix with a line or so change, but making the code
more harder to understand and less structured :-/ Hey, that's life.