MEGA65: making screenshot more robust + code reorg
Especially screenshot-on-exit was affected which only worked in very
special cases, and not in general. This is because taking a screenshot
needs a fully rendered finished frame, but exiting can happen any time.

Thus the new code does "free run" of VIC4 emulation (if screenshot was
requested) in the shutdown handler still hitting end of frame, so the
requested screenshot will be done anyway. This also makes the code at
other places more simple and does not need tricky conseradtions.