MEGA65: HDOS virt. updates/fixes #331
It all began with some patches from @gurcei [THANKS!] to implement some new
HDOS call virtualizations and a fix for cd "..". After trying to merge it, I
realized that I need to modify both of his and mine code. So it ended up a
minor code refactoring/moderniation as well.

Surely more changes will be needed in the future, also more functions
should be virtualized. Initially the scope for HDOS virtualization only
targeted the calls used by the ROM. However it's clear now, that a full
(or almost full) virtualization solution would be much more optimal.