CORE: use SDL_free on SDL allocated ranges
There was a Windows crash reported by ToneDeF on MEGA65 Discord.
It seems the problem that SDL functions allocated and returned a memory
region, should be always free'd with SDL_free(), usual free() and
SDL_free() are not just merely an alias for each other. At least not on
Windows, where it seems, a DLL (SDL2.dll in our case) has their own heap
separated from the program, as kibo kindly descibred.

Though this fix for the "auto-typing" crash on Windows can be cured with
a single line of change, I though I would go deeper, as it's hard to
tell at the "leaf" function, wich kind of memory it has, should be
free'd by which function. Also some review on other usage of SDL_free().