MEGA65: workaround for bad sprite Y in NTSC #428
For unknown reasons, Xemu has 7 (logical raster) off with sprite Y
coordinates in NTSC mode. I can't find the root cause so I applied a
workaround: in NTSC mode an extra xemu-only-internal variable is set
which is applied in sprite rendering. In PAL mode, this var is set to
zero.

Likely, this bug should be fixed in a more sensical way sooner or later.

The problem was originally reported by @MirageBD thanks!