MEGA65: mega65-core fetch/build for 0.97 #436
The reason for the automatic clone/build from mega65-core: it's a bit unfortunate
that over time it's getting harder to understand (even for me) how the previous
hyppo & its company was integrated because I may not have that mega65-core source
tree available anymore with my own tricks to build it, and also some patches
applied there just for Xemu. So my idea here, to automatize the build of the
needed files from mega65-core, so it's a process which can be repeated any time,
without almost any effort, if/when needed for some reason.