MEGA65: hyppo 0.97 for Xemu #436
Merging the result of the `hypup` branch. The previous HYPPO inside Xemu is
about 3 years old for now, which already causes many problems, including the
problems with newest ROMs using the new D81 attach method, just to name one
problem.

Besides this, trying to make GEOS65 a smoother experience with Xemu, as
it does some tricky stuff to be able to work, which caused problems for
Xemu in the past.

Thanks to @gurcei for some preliminary testings in his own repository.